refactor text ident generation
