+++ src/dragtor/llm/llm.py
@@ -200,22 +200,3 @@ class LlamaServerHandler:
     def from_config(cls) -> Self:
         modelpath = config.conf.select("model.file_path", default=None)
         return cls(modelpath)
-
-
-# @dataclass
-# class FeatureGenerator:
-#     llm: LlamaServerHandler = field(default_factory=LlamaServerHandler.from_config)
-#     index: Index = field(default_factory=get_index)
-#
-#     summary_prompt_template: str = """You are an """
-#     def generate_topics(self):
-#         loader = data.JinaLoader()
-#         full_texts = loader.get_cache()
-#
-#         for text in full_texts:
-#             text_id = hashlib.md5(text.encode("utf-8")).hexdigest()
-#
-#             prompt = summary_prompt_template.format(text)
-#
-#             answer = self.llm.query_llm(prompt)
-#             answer_js = json.loads(answer)
